+++ src/proxy/settings_manager.h
@@ -47,7 +47,7 @@ class SettingsManager : public common::patterns::Singleton<SettingsManager> {
   bool GetSendStatistic() const;
   void SetSendStatistic(bool val);
 
-  quint64 GetExecCount() const;
+  uint32_t GetExecCount() const;
 
   void SetDefaultView(supportedViews view);
   supportedViews GetDefaultView() const;
@@ -127,7 +127,7 @@ class SettingsManager : public common::patterns::Singleton<SettingsManager> {
   bool auto_open_console_;
   bool auto_connect_db_;
   bool fast_view_keys_;
-  quint64 exec_count_;
+  uint32_t exec_count_;
 };
 
 }  // namespace proxy
